# docs/api/public-api.openapi.yaml
openapi: 3.0.3
info:
  title: OpenGlass Public API
  version: 0.1.0
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
  description: |
    REST contract for the OpenGlass public layer (MVP scope).

    Source of truth for both `MockApiClient` and the Go backend — see
    `docs/dev/frontend-first-development.md`. No invented wire fields:
    additions require amending this file first.

    Conventions:
    - All endpoints are under `/api/v1`.
    - Invite issuance is an **ops-side operation** (server CLI / admin
      tooling on the instance host), deliberately not part of the public
      client API — there is no "create invite" endpoint here.
    - Auth uses `Authorization: Bearer <access_token>` (JWT, short-lived)
      plus a refresh token cookie/body flow.
    - Error envelope and codes: `docs/api/errors.md`.
    - Realtime events are out of scope here — see `docs/api/ws-events.md`.
servers:
  - url: /api/v1
...
paths:
...
  /chats/{chatId}/attachments:
    post:
      tags: [messages]
      summary: Upload an attachment (photo/file)
      operationId: uploadAttachment
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ChatId' }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [attachment]
                properties:
                  attachment: { $ref: '#/components/schemas/Attachment' }
        '413':
          description: File too large
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /attachments/{attachmentId}:
    get:
      tags: [messages]
      summary: Download an attachment blob
      description: |
        Authorized URL referenced by `Attachment.url`. Serving requires
        membership in the attachment's chat; non-members get 404 so the
        existence of an attachment is never revealed.
      operationId: downloadAttachment
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/AttachmentId' }
      responses:
        '200':
          description: The blob
          content:
            application/octet-stream: {}
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ChatId:
      name: chatId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    AttachmentId:
      name: attachmentId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Missing/expired token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: unauthorized, message: "Token expired" } }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
    Error:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
...
    Attachment:
      type: object
      required: [id, kind, mimeType, sizeBytes, url]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [photo, file] }
        mimeType: { type: string }
        fileName: { type: string }
        sizeBytes: { type: integer }
        url:
          type: string
          description: Authorized URL — requires the same bearer token,
            not a public link (server-trusted, not public internet)
        thumbnailUrl: { type: string, nullable: true }
...